Move widgets to new kas-widgets crate
