+++ crates/kas-widgets/COPYRIGHT
@@ -1,0 +1,6 @@
+This work, the KAS project including the kas and kas-macros libraries,
+is copyrighted by the following contributors:
+
+Diggory Hardy <git@dhardy.name>
+
+This list may be incomplete.
