+++ crates/kas-widgets/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+name = "kas-widgets"
+version = "0.9.1"
+authors = ["Diggory Hardy <git@dhardy.name>"]
+edition = "2018"
+license = "Apache-2.0"
+description = "KAS GUI / widgets"
+readme = "README.md"
+documentation = "https://docs.rs/kas-widgets/"
+keywords = ["gui"]
+categories = ["gui"]
+repository = "https://github.com/kas-gui/kas"
+exclude = ["/screenshots"]
+
+[package.metadata.docs.rs]
+features = ["nightly", "winit", "markdown"]
+rustdoc-args = ["--cfg", "doc_cfg"]
+# To build locally:
+# RUSTDOCFLAGS="--cfg doc_cfg" cargo +nightly doc --features=canvas,svg --no-deps --open
+
+[features]
+# Use min_specialization (enables accelerator underlining for AccelLabel)
+min_spec = []
+
+# Support canvas widget
+canvas = ["tiny-skia"]
+
+# Support SVG images
+svg = ["tiny-skia", "resvg", "usvg"]
+
+[dependencies]
+log = "0.4"
+smallvec = "1.6.1"
+unicode-segmentation = "1.7"
+linear-map = "1.2.0"
+thiserror = "1.0.23"
+image = "0.23.14"
+tiny-skia = { version = "0.5.1", optional = true }
+resvg = { version = "0.15.0", optional = true }
+usvg = { version = "0.15.0", optional = true }
+kas-macros = { version = "0.9.1", path = "../kas-macros" }
+
+# We must rename this package since macros expect kas to be in scope:
+kas = { version = "0.9.1", package = "kas-core", path = "../kas-core" }
