+++ crates/kas-widgets/README.md
@@ -1,0 +1,18 @@
+KAS Widgets
+======
+
+This is KAS's widget library.
+
+For documentation of feature flags, see [Cargo.toml](Cargo.toml).
+
+
+Copyright and Licence
+-------
+
+The [COPYRIGHT](COPYRIGHT) file includes a list of contributors who claim
+copyright on this project. This list may be incomplete; new contributors may
+optionally add themselves to this list.
+
+The KAS library is published under the terms of the Apache License, Version 2.0.
+You may obtain a copy of this licence from the [LICENSE](LICENSE) file or on
+the following webpage: <https://www.apache.org/licenses/LICENSE-2.0>
