+++ crates/kas-widgets/src/adapter/mod.rs
