+++ crates/kas-widgets/src/splitter.rs
