+++ crates/kas-widgets/src/stack.rs
