+++ crates/kas-widgets/src/svg.rs
