+++ crates/kas-widgets/src/view/single_view.rs
